requirements.txt
openai-whisper
torch
moviepy
transformers
sentencepiece
